requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
python-snap7==1.0.0